Python/crossfire/pyproject.toml
[tool.poetry]
name = "crossfire"
version = "0.2.0"
description = "crossfire: Download spatial data sets from crossfire project"
authors = ["Felipe Barros <felipe.b4rros@gmail.com>"]
readme = "README.md"
homepage = "https://fogocruzado.org.br/"
repository = "https://github.com/voltdatalab/crossfire"
classifiers = [
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Visualization",
    "Programming Language :: Python",
]
exclude =  ["man/", "R/", "tables/", "crossfire.Rproj", "DESCRIPTION", "env-example", "Introducao_crossfire.md",
    "Introduction_crossfire.md", "NAMESPACE",]

[tool.poetry.dependencies]
python = "^3.9"
python-decouple = "^3.5"
requests = "^2.26.0"
python-dateutil = "^2.8.2"
DateTime = "^4.3"
pandas = "^1.3.3"
geopandas = "^0.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
